pyproject.toml
[tool.poetry]
name = "voice-recognition-error-anomaly-detection"
version = "0.1.0"
description = ""
authors = ["Jinyamyzk <63580543+Jinyamyzk@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
transformers = "^4.25.1"
fugashi = "^1.2.1"
ipadic = "^1.0.0"
pykakasi = "^2.2.1"
simstring-pure = "^1.0.0"
pandas = "^1.5.2"
openpyxl = "^3.0.10"
mecab-python3 = "^1.0.6"
torchinfo = "^1.7.1"
torch = "~1.9"
torchtext = "~0.10"
scikit-learn = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
